Refactor most of the queue away
