+++ docs/bugfixes.md
@@ -371,11 +371,11 @@ Runic Matrices which are too stable will not fly far out from the center of the 
 
 Fixes a bug where some GUIs would render the wrong aspects when shifting over a slot. Also improves the performance of this overlay.
 
-## Fix Node Rendering Behind Transparent Blocks
+## Fix Custom Renders Rendering Behind Transparent Blocks
 
 **Config option:** `fixNodeRendering`
 
-Fixes a bug where Nodes, Energized Nodes, Node Aspect Tags, and Node Drain Beams would render as though behind transparent blocks when they are in front of them.
+Fixes a bug where Nodes, Energized Nodes, Node Drain Beams, and Aspect Tags, would render as though behind transparent blocks when they are in front of them.
 ## Fix Arcane Bores look-up frequency
 
 **Config option:** `boreDecreaseCVisCheckFrequency`
