+++ src/main/java/dev/rndmorris/salisarcana/mixins/late/thaumcraft/common/tiles/MixinTileNode_FixRenderingLayers.java
@@ -1,0 +1,16 @@
+package dev.rndmorris.salisarcana.mixins.late.thaumcraft.common.tiles;
+
+import net.minecraft.tileentity.TileEntity;
+
+import org.spongepowered.asm.mixin.Mixin;
+
+import thaumcraft.common.tiles.TileEldritchPortal;
+
+@Mixin(TileEldritchPortal.class)
+public abstract class MixinTileNode_FixRenderingLayers extends TileEntity {
+
+    @Override
+    public boolean shouldRenderInPass(int pass) {
+        return pass == 1;
+    }
+}
